[bp] Add transitive dependencies with forbidden access to classpath
   The Java compiler needs to resolve types from transitive dependencies
   (e.g. when a superclass or interface comes from a bundle not directly
   required by the current project). Add all transitive requirements
   with forbidden access rules to ensure compilation succeeds while
   preventing direct use of types from not imported bundles.

This mirrors PDE's RequiredPluginsClasspathContainer behavior added in
   eclipse-pde/eclipse.pde#2218.

Add integration test for transitive dependency classpath resolution

   Verifies that transitive dependencies are added to the compiler
   classpath with forbidden access rules, ensuring compilation succeeds
   when types from indirectly required bundles need to be resolved
   (e.g. forEach() via an intermediate Iterable interface).
