+++ tycho-its/projects/compiler.transitiveDependency/middle.bundle/META-INF/MANIFEST.MF
@@ -1,0 +1,8 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Middle Bundle
+Bundle-SymbolicName: middle.bundle
+Bundle-Version: 1.0.0.qualifier
+Export-Package: middle.bundle
+Import-Package: base.bundle
+Bundle-RequiredExecutionEnvironment: JavaSE-17
