+++ tycho-its/projects/compiler.transitiveDependency/middle.bundle/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
