+++ tycho-its/src/test/java/org/eclipse/tycho/test/compiler/CompilerClasspathTest.java
@@ -25,4 +25,19 @@ public class CompilerClasspathTest extends AbstractTychoIntegrationTest {
 		verifier.verifyErrorFreeLog();
 	}
 
+	/**
+	 * Tests that transitive dependencies are added to the compiler classpath with
+	 * forbidden access rules. This is needed when a bundle imports a package from
+	 * another bundle whose types extend/implement types from a third bundle that is
+	 * not directly imported. Without transitive deps on the classpath, the compiler
+	 * cannot resolve the type hierarchy (e.g. forEach() from Iterable via an
+	 * intermediate interface).
+	 */
+	@Test
+	public void testTransitiveDependencyOnClasspath() throws Exception {
+		Verifier verifier = getVerifier("compiler.transitiveDependency", false);
+		verifier.executeGoal("compile");
+		verifier.verifyErrorFreeLog();
+	}
+
 }
